requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
python-multipart==0.0.20
pydantic==2.9.2
eval7==0.1.10
python-dotenv==1.0.1
azure-storage-blob==12.19.1